WIP: Migrate :screen-piechart to ComposeViewModel
